+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Area outside gameboard filled.
 - Respawn player on space key down.
 - Respawning on site of demise rather than in center of gameboard.
+- Replace dead blobs.
 
 ## [0.6.0] - 2016-12-19
 ### Added